custom hook
Created custom hook useSignup that validates for possible errors outside Firebase's default validations. It creates a user with given email and password and sets property displayName of the returned object equal to the input received in the form in the Sigunup.js component. No restrictions added on the displayName yet